+++ apps/desktop-wallet/src/modals/WalletConnect/SignUnsignedTxModal.tsx
